Refactored SystemData into SystemData + SystemManager.
Removed ScraperCmdLine.h/.cpp, since it was never used anyway.
