--- es-app/src/ScraperCmdLine.h
@@ -1,3 +1,0 @@
-#pragma once
-
-int run_scraper_cmdline();
