+++ es-app/src/guis/GuiScraperMulti.cpp
@@ -2,6 +2,7 @@
 #include "Renderer.h"
 #include "Log.h"
 #include "views/ViewController.h"
+#include "SystemManager.h"
 
 #include "components/TextComponent.h"
 #include "components/ButtonComponent.h"
@@ -71,7 +72,7 @@ GuiScraperMulti::GuiScraperMulti(Window* window, const std::queue<ScraperSearchP
 GuiScraperMulti::~GuiScraperMulti()
 {
 	// view type probably changed (basic -> detailed)
-	for(auto it = SystemData::sSystemVector.begin(); it != SystemData::sSystemVector.end(); it++)
+	for(auto it = SystemManager::getInstance()->getSystems().begin(); it != SystemManager::getInstance()->getSystems().end(); it++)
 		ViewController::get()->reloadGameListView(*it, false);
 }
