+++ es-app/src/views/SystemView.cpp
@@ -7,6 +7,7 @@
 #include "animations/LambdaAnimation.h"
 #include "SystemData.h"
 #include "Settings.h"
+#include "SystemManager.h"
 #include "Util.h"
 
 #define SELECTED_SCALE 1.5f
@@ -32,7 +33,8 @@ void SystemView::populate()
 {
 	mEntries.clear();
 
-	for(auto it = SystemData::sSystemVector.begin(); it != SystemData::sSystemVector.end(); it++)
+	const std::vector<SystemData*>& systems = SystemManager::getInstance()->getSystems();
+	for(auto it = systems.begin(); it != systems.end(); it++)
 	{
 		const std::shared_ptr<ThemeData>& theme = (*it)->getTheme();
